Switch to package names without user scope (do not propose to upstream)
The next commit contains substantial changes.  It is unlikely that we
will send them as a pull request right away.  Switch to package names
without user scope to emphasize that the package will probably not be
published to Atmosphere in its present form and must be installed by
cloning Git.

This commit must not be proposed to upstream.

[spr: Split rename and package.js changes into two commits.]

Signed-off-by: Steffen Prohaska <prohaska@zib.de>